combining test methods, removing redundant
